Fixed bugs in the updater
Fixed a couple of bugs in the updater that would cause endless loop of starting the ps3mfw builder.
the bug when the internet was out and could not connect to server and proggram would hang instead of closing.
the bug when the version.txt or ps3mfw.exe files did not exist. Proceeded in a check of the files existence and if not create them and download the proggram if it isn't directly from the updater!